completed the left overs

+++ controllers/advocateController.js
@@ -1,7 +1,10 @@
 const mongoose = require("mongoose");
 const Advocate = require("../models/advocateModel");
+const Review = require("../models/reviewModel");
 const catchAsync = require("../middlewares/catchAsync");
 const ExpressError = require("../middlewares/ExpressError");
+const { message } = require("../models/JOI/reviewJOISchema");
+const { storeReturnTo } = require("../middlewares/isLoggedIn");
 
 // @route GET advocates/all
 // @desc shows all advocates
@@ -16,7 +19,15 @@ const getAllAdvocates = catchAsync(async (req, res) => {
 // @access PUBLIC
 const getAdvocate = catchAsync(async (req, res) => {
     const { id } = req.params;
-    const advocate = await Advocate.findById(id);
+    const advocate = await Advocate.findById(id)
+        .populate({
+            path: "reviews",
+            populate: {
+                path: "author",
+            },
+        })
+        .populate("author");
+    console.log(advocate);
     res.status(200).render("advocate/show", { advocate });
 });
 
@@ -24,11 +35,16 @@ const getAdvocate = catchAsync(async (req, res) => {
 // @desc register new Advocate
 // @access PUBLIC
 const addAdvocate = catchAsync((req, res) => {
+    if (!req.isAuthenticated()) {
+        req.session.returnTo = req.originalUrl;
+        throw new ExpressError("You need to be logged in to add a profile");
+    }
     res.render("advocate/add");
 });
 
 const saveNewAdvocate = catchAsync(async (req, res) => {
     const advocate = new Advocate(req.body.advocate);
+    advocate.author = req.User._id;
     await advocate.save();
     res.redirect(`/advocates/${advocate._id}`);
 });
@@ -37,6 +53,10 @@ const saveNewAdvocate = catchAsync(async (req, res) => {
 // @desc rewite an advocate's details
 // @access PUBLIC
 const editAdvocateProfile = catchAsync(async (req, res) => {
+    if (!req.isAuthenticated()) {
+        req.session.returnTo = req.originalUrl;
+        throw new ExpressError("You need to be logged in to add a profile");
+    }
     const { id } = req.params;
     const advocate = await Advocate.findById(id);
     res.status(200).render("advocate/edit", { advocate });
@@ -53,10 +73,46 @@ const reSaveAdvocateProfile = catchAsync(async (req, res) => {
 // @desc delete an advocate's profile
 // @access PUBLIC
 const deleteAdvocateProfile = catchAsync(async (req, res) => {
+    if (!req.isAuthenticated()) {
+        req.session.returnTo = req.originalUrl;
+        throw new ExpressError("You need to be logged in to add a profile");
+    }
     const { id } = req.params;
     const advocate = await Advocate.findById(id);
     await Advocate.findByIdAndDelete(id);
     res.status(200).redirect("/advocates/all");
+});
+
+// @route POST advocates/:id/
+// @desc give an advocate a review
+// @access PUBLIC
+const giveReview = catchAsync(async (req, res) => {
+    if (!req.isAuthenticated()) {
+        req.session.returnTo = req.originalUrl;
+        throw new ExpressError("You need to be logged in to add a profile");
+    }
+    const advocate = await Advocate.findById(req.params.id);
+    const review = new Review(req.body.review);
+    review.author = req.user._id;
+    advocate.reviews.push(review);
+    await review.save();
+    await advocate.save();
+    res.redirect(`/advocates/${advocate._id}`);
+});
+
+// @route DELETE advocates/:id/reviews/:reviewId
+// @desc delete an advocate's review
+// @access PUBLIC
+const removeReview = catchAsync(async (req, res) => {
+    if (!req.isAuthenticated()) {
+        req.session.returnTo = req.originalUrl;
+        throw new ExpressError("You need to be logged in to add a profile");
+    }
+    const { id, reviewId } = req.params;
+    console.log(id, reviewId);
+    await Advocate.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
+    await Review.findByIdAndDelete(reviewId);
+    res.redirect(`/advocates/${id}`);
 });
 
 module.exports = {
@@ -67,4 +123,6 @@ module.exports = {
     editAdvocateProfile,
     reSaveAdvocateProfile,
     deleteAdvocateProfile,
+    giveReview,
+    removeReview,
 };
